Worked on the jump ball code. Also switched a lot of variables from int to size_t. Changed playerWithBall to playerWithBallInstance or playerWithBallID depending on what was appropriate. Fixed the android build system to use gcc-4.9 since Aide now uses that instead of 4.6

+++ include/jumpballs.h
@@ -44,8 +44,8 @@ class jumpBalls
     teamTypes getBallTippedToTeam();  // retrieves the value of the ballTippedToTeam
     void setBallTippedToTeam(teamTypes set);	 // sets the value of the ballTippedToTeam
 
-    int getBallTippedToPlayerID();  // retrieves the value of the ballTippedToPlayerID
-    void setBallTippedToPlayerID(int player);  // sets the value of the ballTippedToPlayerID
+    size_t getBallTippedToPlayerID();  // retrieves the value of the ballTippedToPlayerID
+    void setBallTippedToPlayerID(size_t player);  // sets the value of the ballTippedToPlayerID
 
     playerPositions getBallTippedToPosition();  // retrieves the value of ballTippedToPosition
     void setBallTippedToPosition(playerPositions set);  // sets the value of ballTippedToPosition
